+++ BambooTracker/gui/color_palette.cpp
@@ -12,10 +12,10 @@ ColorPalette::ColorPalette()
 	// Instrument editor
 	instFMEnvLine1Color = QColor::fromRgb(242, 38, 19, 255);
 	instFMEnvLine2Color = QColor::fromRgb(46, 204, 113, 255);
+	instFMEnvLine3Color = QColor::fromRgb(38, 38, 255, 255);
+	instFMEnvGridColor = QColor::fromRgb(40, 40, 40, 63);
 	instFMEnvBackColor = QColor::fromRgb(255, 255, 255, 0);
 	instFMEnvBorderColor = QColor::fromRgb(125, 125, 125, 255);
-	instFMEnvCirclePenColor = QColor::fromRgb(0, 0, 0, 255);
-	instFMEnvCircleBrushColor = QColor::fromRgb(255, 255, 255, 255);
 	instFMAlForeColor = QColor::fromRgb(107, 185, 240, 255);
 	instFMAlBackColor = QColor::fromRgb(255, 255, 255, 0);
 	instSeqLoopBackColor = QColor::fromRgb(25, 25, 25, 255);