+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+### Changed
+- Change drawing method of FM envelope in FM instrument editor
+
 ### Fixed
 - Fix incorrect text painting in the pattern editor
 - [#168] - Fix pattern rendering glitches (thanks [@OPNA2608], [@YuzuMSX])